Update Privacy settings and move it to MainSettings page
