+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/settings/about/AboutFragment.kt
@@ -20,14 +20,12 @@ import io.horizontalsystems.bankwallet.R
 import io.horizontalsystems.bankwallet.core.BaseComposeFragment
 import io.horizontalsystems.bankwallet.core.composablePage
 import io.horizontalsystems.bankwallet.core.composablePopup
-import io.horizontalsystems.bankwallet.core.slideFromBottom
 import io.horizontalsystems.bankwallet.core.stats.StatEvent
 import io.horizontalsystems.bankwallet.core.stats.StatPage
 import io.horizontalsystems.bankwallet.core.stats.stat
 import io.horizontalsystems.bankwallet.modules.releasenotes.ReleaseNotesScreen
 import io.horizontalsystems.bankwallet.modules.settings.appstatus.AppStatusScreen
 import io.horizontalsystems.bankwallet.modules.settings.main.HsSettingCell
-import io.horizontalsystems.bankwallet.modules.settings.privacy.PrivacyScreen
 import io.horizontalsystems.bankwallet.modules.settings.terms.TermsScreen
 import io.horizontalsystems.bankwallet.ui.compose.ComposeAppTheme
 import io.horizontalsystems.bankwallet.ui.compose.components.AppBar
@@ -48,7 +46,6 @@ class AboutFragment : BaseComposeFragment() {
 private const val AboutPage = "about"
 private const val ReleaseNotesPage = "release_notes"
 private const val AppStatusPage = "app_status"
-private const val PrivacyPage = "privacy"
 private const val TermsPage = "terms"
 
 @Composable
@@ -61,7 +58,6 @@ private fun AboutNavHost(fragmentNavController: NavController) {
         composable(AboutPage) {
             AboutScreen(
                 navController,
-                { fragmentNavController.slideFromBottom(R.id.contactOptionsDialog) },
                 { fragmentNavController.popBackStack() }
             )
         }
@@ -69,7 +65,6 @@ private fun AboutNavHost(fragmentNavController: NavController) {
             ReleaseNotesScreen(false, { navController.popBackStack() })
         }
         composablePage(AppStatusPage) { AppStatusScreen(navController) }
-        composablePage(PrivacyPage) { PrivacyScreen(navController) }
         composablePopup(TermsPage) { TermsScreen(navController) }
     }
 }
@@ -77,7 +72,6 @@ private fun AboutNavHost(fragmentNavController: NavController) {
 @Composable
 private fun AboutScreen(
     navController: NavController,
-    showContactOptions: () -> Unit,
     onBackPress: () -> Unit,
     aboutViewModel: AboutViewModel = viewModel(factory = AboutModule.Factory()),
 ) {
@@ -147,16 +141,6 @@ private fun SettingSections(
                     stat(page = StatPage.AboutApp, event = StatEvent.Open(StatPage.Terms))
                 }
             )
-        }, {
-            HsSettingCell(
-                R.string.Settings_Privacy,
-                R.drawable.ic_user_20,
-                onClick = {
-                    navController.navigate(PrivacyPage)
-
-                    stat(page = StatPage.AboutApp, event = StatEvent.Open(StatPage.Privacy))
-                }
-            )
         })
     )
 
@@ -170,7 +154,7 @@ private fun SettingSections(
                 onClick = {
                     LinkHelper.openLinkInAppBrowser(context, viewModel.githubLink)
 
-                    stat(page = StatPage.AboutApp, event= StatEvent.Open(StatPage.ExternalGithub))
+                    stat(page = StatPage.AboutApp, event = StatEvent.Open(StatPage.ExternalGithub))
                 }
             )
         }, {
@@ -180,11 +164,11 @@ private fun SettingSections(
                 onClick = {
                     LinkHelper.openLinkInAppBrowser(context, viewModel.appWebPageLink)
 
-                    stat(page = StatPage.AboutApp, event= StatEvent.Open(StatPage.ExternalWebsite))
+                    stat(page = StatPage.AboutApp, event = StatEvent.Open(StatPage.ExternalWebsite))
                 }
             )
         })
     )
-    
+
     VSpacer(32.dp)
 }
