+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/settings/privacy/PrivacySettingsFragment.kt
@@ -22,29 +22,111 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
+import androidx.fragment.app.viewModels
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
 import io.horizontalsystems.bankwallet.R
+import io.horizontalsystems.bankwallet.core.BaseComposeFragment
+import io.horizontalsystems.bankwallet.core.paidAction
 import io.horizontalsystems.bankwallet.core.stats.StatEvent
 import io.horizontalsystems.bankwallet.core.stats.StatPage
 import io.horizontalsystems.bankwallet.core.stats.stat
+import io.horizontalsystems.bankwallet.modules.main.MainModule
+import io.horizontalsystems.bankwallet.modules.settings.privacy.tor.SecurityTorSettingsModule
+import io.horizontalsystems.bankwallet.modules.settings.privacy.tor.SecurityTorSettingsViewModel
+import io.horizontalsystems.bankwallet.modules.settings.security.ui.TorBlock
 import io.horizontalsystems.bankwallet.ui.compose.ComposeAppTheme
 import io.horizontalsystems.bankwallet.ui.compose.components.AppBar
 import io.horizontalsystems.bankwallet.ui.compose.components.HFillSpacer
 import io.horizontalsystems.bankwallet.ui.compose.components.HSpacer
 import io.horizontalsystems.bankwallet.ui.compose.components.HsBackButton
 import io.horizontalsystems.bankwallet.ui.compose.components.HsSwitch
-import io.horizontalsystems.bankwallet.ui.compose.components.InfoTextBody
+import io.horizontalsystems.bankwallet.ui.compose.components.InfoText
+import io.horizontalsystems.bankwallet.ui.compose.components.PremiumHeader
+import io.horizontalsystems.bankwallet.ui.compose.components.TextImportantWarning
 import io.horizontalsystems.bankwallet.ui.compose.components.VSpacer
 import io.horizontalsystems.bankwallet.ui.compose.components.body_leah
 import io.horizontalsystems.bankwallet.ui.compose.components.cell.CellUniversal
-import io.horizontalsystems.bankwallet.ui.compose.components.cell.SectionUniversalLawrence
+import io.horizontalsystems.bankwallet.ui.compose.components.cell.SectionPremiumUniversalLawrence
+import io.horizontalsystems.bankwallet.ui.extensions.ConfirmationDialog
+import io.horizontalsystems.subscriptions.core.PrivacyMode
+import kotlin.system.exitProcess
+
+class PrivacySettingsFragment : BaseComposeFragment() {
+
+    private val torViewModel by viewModels<SecurityTorSettingsViewModel> {
+        SecurityTorSettingsModule.Factory()
+    }
+
+    @Composable
+    override fun GetContent(navController: NavController) {
+        PrivacyScreen(
+            navController = navController,
+            torViewModel = torViewModel,
+            showAppRestartAlert = { showAppRestartAlert() },
+            restartApp = { restartApp() },
+        )
+    }
+
+    private fun showAppRestartAlert() {
+        val warningTitle = if (torViewModel.torCheckEnabled) {
+            getString(R.string.Tor_Connection_Enable)
+        } else {
+            getString(R.string.Tor_Connection_Disable)
+        }
+
+        val actionButton = if (torViewModel.torCheckEnabled) {
+            getString(R.string.Button_Enable)
+        } else {
+            getString(R.string.Button_Disable)
+        }
+
+        ConfirmationDialog.show(
+            icon = R.drawable.ic_tor_connection_24,
+            title = getString(R.string.Tor_Alert_Title),
+            warningTitle = warningTitle,
+            warningText = getString(R.string.SettingsSecurity_AppRestartWarning),
+            actionButtonTitle = actionButton,
+            transparentButtonTitle = getString(R.string.Alert_Cancel),
+            fragmentManager = childFragmentManager,
+            listener = object : ConfirmationDialog.Listener {
+                override fun onActionButtonClick() {
+                    torViewModel.setTorEnabled()
+                }
+
+                override fun onTransparentButtonClick() {
+                    torViewModel.resetSwitch()
+                }
+
+                override fun onCancelButtonClick() {
+                    torViewModel.resetSwitch()
+                }
+            }
+        )
+    }
+
+    private fun restartApp() {
+        activity?.let {
+            MainModule.startAsNewTask(it)
+            exitProcess(0)
+        }
+    }
+}
 
 @Composable
-fun PrivacyScreen(navController: NavController) {
+fun PrivacyScreen(
+    navController: NavController,
+    torViewModel: SecurityTorSettingsViewModel,
+    showAppRestartAlert: () -> Unit = {},
+    restartApp: () -> Unit = {},
+) {
     val viewModel = viewModel<PrivacyViewModel>(factory = PrivacyViewModel.Factory())
-
     val uiState = viewModel.uiState
+
+    if (torViewModel.restartApp) {
+        restartApp()
+        torViewModel.appRestarted()
+    }
 
     Column(
         modifier = Modifier
@@ -62,20 +144,20 @@ fun PrivacyScreen(navController: NavController) {
                 .weight(1f)
                 .verticalScroll(rememberScrollState())
         ) {
-            InfoTextBody(
+            VSpacer(12.dp)
+            TextImportantWarning(
+                modifier = Modifier.padding(horizontal = 16.dp),
                 text = stringResource(R.string.Privacy_Information),
             )
 
-            BulletedText(R.string.Privacy_BulletedText1)
-            BulletedText(R.string.Privacy_BulletedText2)
-            BulletedText(R.string.Privacy_BulletedText3)
-            VSpacer(height = 16.dp)
-            SectionUniversalLawrence {
+            VSpacer(height = 24.dp)
+            PremiumHeader()
+            SectionPremiumUniversalLawrence {
                 CellUniversal {
                     Icon(
                         painter = painterResource(id = R.drawable.ic_share_24px),
                         contentDescription = "Share",
-                        tint = ComposeAppTheme.colors.grey
+                        tint = ComposeAppTheme.colors.jacob
                     )
                     HSpacer(width = 16.dp)
                     body_leah(text = stringResource(R.string.ShareUiData))
@@ -83,13 +165,30 @@ fun PrivacyScreen(navController: NavController) {
                     HsSwitch(
                         checked = uiState.uiStatsEnabled,
                         onCheckedChange = {
-                            viewModel.toggleUiStats(it)
+                            navController.paidAction(PrivacyMode) {
+                                viewModel.toggleUiStats(it)
 
-                            stat(page = StatPage.Privacy, event = StatEvent.EnableUiStats(it))
+                                stat(page = StatPage.Privacy, event = StatEvent.EnableUiStats(it))
+                            }
                         }
                     )
                 }
             }
+            InfoText(
+                text = stringResource(R.string.ShareUiDataDescription),
+            )
+
+            VSpacer(20.dp)
+            SectionPremiumUniversalLawrence {
+                TorBlock(
+                    torViewModel,
+                    navController,
+                    showAppRestartAlert,
+                )
+            }
+            InfoText(
+                text = stringResource(R.string.SettingsSecurity_TorConnectionDescription),
+            )
         }
 
         Divider(
