+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/settings/security/ui/TorBlock.kt
@@ -11,12 +11,11 @@ import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import io.horizontalsystems.bankwallet.R
 import io.horizontalsystems.bankwallet.core.paidAction
+import io.horizontalsystems.bankwallet.modules.settings.privacy.tor.SecurityTorSettingsViewModel
 import io.horizontalsystems.bankwallet.modules.settings.security.SecurityCenterCell
-import io.horizontalsystems.bankwallet.modules.settings.security.tor.SecurityTorSettingsViewModel
 import io.horizontalsystems.bankwallet.ui.compose.ComposeAppTheme
 import io.horizontalsystems.bankwallet.ui.compose.components.HsSwitch
 import io.horizontalsystems.bankwallet.ui.compose.components.body_leah
-import io.horizontalsystems.bankwallet.ui.compose.components.cell.SectionPremiumUniversalLawrence
 import io.horizontalsystems.subscriptions.core.Tor
 
 @Composable
@@ -30,33 +29,31 @@ fun TorBlock(
         viewModel.restartAppAlertShown()
     }
 
-    SectionPremiumUniversalLawrence {
-        SecurityCenterCell(
-            start = {
-                Icon(
-                    modifier = Modifier.size(24.dp),
-                    painter = painterResource(R.drawable.ic_tor_connection_24),
-                    tint = ComposeAppTheme.colors.jacob,
-                    contentDescription = null,
-                )
-            },
-            center = {
-                body_leah(
-                    text = stringResource(R.string.Tor_Title),
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                )
-            },
-            end = {
-                HsSwitch(
-                    checked = viewModel.torCheckEnabled,
-                    onCheckedChange = { checked ->
-                        navController.paidAction(Tor) {
-                            viewModel.setTorEnabledWithChecks(checked)
-                        }
+    SecurityCenterCell(
+        start = {
+            Icon(
+                modifier = Modifier.size(24.dp),
+                painter = painterResource(R.drawable.ic_tor_connection_24),
+                tint = ComposeAppTheme.colors.jacob,
+                contentDescription = null,
+            )
+        },
+        center = {
+            body_leah(
+                text = stringResource(R.string.Tor_Title),
+                maxLines = 1,
+                overflow = TextOverflow.Ellipsis,
+            )
+        },
+        end = {
+            HsSwitch(
+                checked = viewModel.torCheckEnabled,
+                onCheckedChange = { checked ->
+                    navController.paidAction(Tor) {
+                        viewModel.setTorEnabledWithChecks(checked)
                     }
-                )
-            }
-        )
-    }
+                }
+            )
+        }
+    )
 }
