+++ app/src/main/res/navigation/main_graph.xml
@@ -83,6 +83,9 @@
         android:id="@+id/securitySettingsFragment"
         android:name="io.horizontalsystems.bankwallet.modules.settings.security.SecuritySettingsFragment"/>
     <fragment
+        android:id="@+id/privacySettingsFragment"
+        android:name="io.horizontalsystems.bankwallet.modules.settings.privacy.PrivacySettingsFragment"/>
+    <fragment
         android:id="@+id/blockchainSettingsFragment"
         android:name="io.horizontalsystems.bankwallet.modules.blockchainsettings.BlockchainSettingsFragment"/>
     <fragment
